+++ CHANGELOG.md
@@ -2,7 +2,9 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2022.06.27`.
  * NOTE: Relaxed processing of Youtube Playlist API response.
+ * NOTE: Store and use verified URL and repository results from pana.
 
 ## `20220623t124000-all`
  * Bumped runtimeVersion to `2022.06.23`.